Fix done

+++ src/com/John/Entity/Item.java
@@ -1,8 +1,11 @@
 package com.John.Entity;
-
+/**
+ * author: Johnson Hartanto
+ * NRP: 1772017
+ */
 public class Item {
     private String name;
-    private int price;
+    private Double price;
     private Category category;
 
     public String getName() {
@@ -13,11 +16,11 @@ public class Item {
         this.name = name;
     }
 
-    public int getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
